+++ R/chop.R
@@ -48,6 +48,7 @@ NULL
 #' `close_end` arguments, and should return an object of class `breaks`.
 #' Use `brk_*` functions to create a variety of data-dependent breaks.
 #'
+#'
 #' ## Options for breaks
 #'
 #' By default, left-closed intervals are created. If `left` is `FALSE`, right-
@@ -92,6 +93,11 @@ NULL
 #' `labels` may be a character vector. It should have the same length as the
 #' number of intervals. Alternatively, use a `lbl_*` function such as
 #' [lbl_seq()].
+#'
+#' If `breaks` is a named vector, and no `labels` argument is explicitly
+#' given, then non-zero-length names of `breaks` will be used as labels.
+#' You can omit the name of the last element. Be careful if elements of `x`
+#' may be below the first break.
 #'
 #' If `labels` is `NULL`, then integer codes will be returned instead of a
 #' factor.
@@ -148,6 +154,12 @@ NULL
 #'
 #' chop(1:10, c(2, 5, 8), labels = lbl_dash())
 #'
+#' chop(1:10, c(Low = 1, Mid = 3, High = 8, 10))
+#'
+#' # Take care when using named breaks:
+#' chop(1:6, c(One_plus = 1, Two_plus = 2, Three_plus = 3, Five_plus = 5))
+#' chop(0:4, c(One_plus = 1, Two_plus = 2, Three_plus = 3, Five_plus = 5))
+#'
 #' # floating point inaccuracy:
 #' chop(0.3/3, c(0, 0.1, 0.1, 1), labels = c("< 0.1", "0.1", "> 0.1"))
 #'
@@ -166,6 +178,7 @@ chop <- function (x, breaks,
           is.flag(drop),
           is.flag(raw) || is.null(raw)
         )
+  break_names <- names(breaks)
   if (! is.function(breaks)) breaks <- brk_default(breaks)
   breaks <- breaks(x, extend, left, close_end)
   assert_that(is.breaks(breaks), length(breaks) >= 2L)
@@ -174,7 +187,9 @@ chop <- function (x, breaks,
 
   if (is.null(labels)) return(codes)
 
-  lbls <- if (is.function(labels)) {
+  lbls <- if (! is.null(break_names) && missing(labels)) {
+    break_names[nzchar(break_names)]
+  } else if (is.function(labels)) {
     if (is.null(raw)) labels(breaks) else labels(breaks, raw = raw)
   } else {
     labels